baseapp: Move query tests to their own file
